Ajustes persistencia Paciente y Terapista

+++ src/main/java/com/poa/tp/repositories/PacienteRepository.java
@@ -4,8 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.poa.tp.entities.Paciente;
 
-
-public interface PacienteRepository extends JpaRepository<Paciente, Integer> {
-
+public interface PacienteRepository extends JpaRepository<Paciente, Long> {
 
 }
