+++ src/main/java/com/poa/tp/repositories/TerapistaRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.poa.tp.entities.Terapista;
 
 
-public interface TerapistaRepository extends JpaRepository<Terapista, Integer> {
+public interface TerapistaRepository extends JpaRepository<Terapista, Long> {
 
 
 }
